Relaxed some Monad constraints to Applicative, fixed tests
